Minor errores fixed and UML updated

+++ src/main/java/acme/entities/trainings/TrainingModule.java
@@ -60,7 +60,7 @@ public class TrainingModule extends AbstractEntity {
 	private String				optionalLink;
 
 	@NotNull
-	private Double				totalTime;
+	private Integer				totalTime;
 
 	// Relationships ----------------------------------------------------------
